Refactor permutation_iterator
